+++ include/compressedSA.hpp
@@ -11,7 +11,7 @@ class compressedSA
 {   
 private:
     std::unordered_map<uint64_t, hashValue> hashMap;
-    std::vector<int> CSA;
+    std::vector<uint64_t> CSA;
     std::string text;
 
 public:
@@ -34,13 +34,13 @@ public:
         }
     uint64_t encode_dna5(const std::string& kmer);
     std::string decode_dna5(uint64_t encoded, unsigned k);
-    std::vector<int> findPattern(std::string& kmer, unsigned k);
+    std::vector<uint64_t> findPattern(std::string& kmer, unsigned k);
     static compressedSA compute (const std::string& fastaData, const unsigned k);
     size_t memoryUsageBytes() const;
 
 
 
-    compressedSA(std::unordered_map<uint64_t, hashValue> hashMap, std::vector<int> CSA, std::string text)
+    compressedSA(std::unordered_map<uint64_t, hashValue> hashMap, std::vector<uint64_t> CSA, std::string text)
     {
         this -> hashMap = hashMap;
         this -> CSA = CSA;
@@ -48,8 +48,10 @@ public:
     }
 
     compressedSA (const std::string& fastaData, const unsigned k)
-    {
+    {   
+        
        *this = compute(fastaData, k);
+       
     }
     ~compressedSA() {}
 };
